move get_image_set
