+++ allensdk/brain_observatory/behavior/behavior_project_cache/tables/util/image_presentation_utils.py
@@ -1,0 +1,33 @@
+from typing import Optional
+import pandas as pd
+import re
+
+
+def get_image_set(df: pd.DataFrame) -> pd.Series:
+    """Get image set
+
+    The image set here is the letter part of the session type
+    ie for session type OPHYS_1_images_B, it would be "B"
+
+    Some session types don't have an image set name, such as
+    gratings, which will be set to null
+
+    Parameters
+    ----------
+    df
+        The session df
+
+    Returns
+    --------
+    Series with index same as df whose values are image_set
+    """
+    def __get_image_set_name(session_type: Optional[str]):
+        match = re.match(r'.*images_(?P<image_set>\w)', session_type)
+        if match is None:
+            return None
+        return match.group('image_set')
+
+    session_type = df['session_type'][
+        df['session_type'].notnull()]
+    image_set = session_type.apply(__get_image_set_name)
+    return image_set
