+++ allensdk/brain_observatory/behavior/behavior_project_cache/tables/util/prior_exposure_processing.py
@@ -4,6 +4,9 @@ from typing import Optional
 import pandas as pd
 
 from allensdk.brain_observatory.behavior.behavior_project_cache.project_apis.data_io import BehaviorProjectLimsApi  # noqa: E501
+from allensdk.brain_observatory.behavior.behavior_project_cache \
+    .tables.util.image_presentation_utils import (
+        get_image_set)
 
 
 def get_prior_exposures_to_session_type(df: pd.DataFrame) -> pd.Series:
@@ -20,36 +23,6 @@ def get_prior_exposures_to_session_type(df: pd.DataFrame) -> pd.Series:
     session type
     """
     return __get_prior_exposure_count(df=df, to=df['session_type'])
-
-
-def get_image_set(df: pd.DataFrame) -> pd.Series:
-    """Get image set
-
-    The image set here is the letter part of the session type
-    ie for session type OPHYS_1_images_B, it would be "B"
-
-    Some session types don't have an image set name, such as
-    gratings, which will be set to null
-
-    Parameters
-    ----------
-    df
-        The session df
-
-    Returns
-    --------
-    Series with index same as df whose values are image_set
-    """
-    def __get_image_set_name(session_type: Optional[str]):
-        match = re.match(r'.*images_(?P<image_set>\w)', session_type)
-        if match is None:
-            return None
-        return match.group('image_set')
-
-    session_type = df['session_type'][
-        df['session_type'].notnull()]
-    image_set = session_type.apply(__get_image_set_name)
-    return image_set
 
 
 def get_prior_exposures_to_image_set(df: pd.DataFrame) -> pd.Series:
